Bids controllers and functional and views improvements!

+++ farmerMarket/app/Http/Controllers/CommentsController.php
@@ -36,6 +36,11 @@ class CommentsController extends Controller
     public function delete($id){
         $com = Comments::findOrFail($id);
 
+        if($com->user_id != Auth::id() || Auth::user()->admin){
+            session()->flash('error','Resource not allowed to you!');
+            return redirect('/');
+        }
+
         if(count($com->hasReplay()) > 0){
             $com->hasReplay()->delete();
         }
